follow strict cast rules in ww code , make another target for lwip with complie options that 'lwip/contribs' gave to it, our options war too strict for lwip

+++ tunnels/TcpConnector/common/freebind.c
@@ -18,12 +18,12 @@ bool tcpconnectorApplyFreeBindRandomDestIp(tunnel_t* t,address_context_t *dest_c
         // no probelm if overflows
         {
 #ifdef OS_UNIX
-            const uint32_t large_random = (((uint32_t) rand_r(&seed)) % state->outbound_ip_range);
+            const uint32_t large_random = (uint32_t)(((uint64_t) rand_r(&seed)) % state->outbound_ip_range);
 #else
-            const uint32_t large_random = (((uint32_t) rand_s(&seed)) % state->outbound_ip_range);
+            const uint32_t large_random = (uint32_t)(((uint64_t) rand_s(&seed)) % state->outbound_ip_range);
 #endif
             uint32_t calc = ntohl((uint32_t) dest_ctx->ip_address.u_addr.ip4.addr);
-            calc          = calc & ~(state->outbound_ip_range - 1ULL);
+            calc          = calc & ~(((uint32_t)state->outbound_ip_range) - 1U);
             calc          = htonl(calc + large_random);
 
             memoryCopy(&(dest_ctx->ip_address.u_addr.ip4), &calc, sizeof(struct in_addr));
