+++ ww/bufio/buffer_queue.c
@@ -75,5 +75,5 @@ sbuf_t *bufferqueuePop(buffer_queue_t *self)
 */
 size_t bufferqueueLen(buffer_queue_t *self)
 {
-    return queue_t_size(&self->q);
+    return (size_t) ( queue_t_size(&self->q));
 }
